mv_fashion/requirements.txt
Pillow
Django==4.0.2
django-phonenumber-field[phonenumbers]
django-crispy-forms